feat(axios): use env variable for api url and increase timeout
The changes in this commit update the `apiUrl` and `timeout` values in the `axios.ts` file.

The `apiUrl` is now determined based on the `NODE_ENV` environment variable. If the environment is "production", the `LARAVEL_API_URL` environment variable is used. Otherwise, the default URL is set to "http://127.0.0.1:8000/api/v1".

The `timeout` value for the Axios instance has been increased from 15 seconds to 25 seconds to allow for longer-running API requests.

These changes ensure that the API URL and timeout are properly configured for both development and production environments.