feat(cut-dialog): Fetch additional cut lists when needed
When the user opens the cut dialog, we check if the current number of cut
lists is less than 10. If so, we fetch an additional 20 cut lists to ensure
the user has a sufficient number of cut lists to choose from. This
improvement enhances the user experience by providing more options and
reducing the need for the user to manually fetch more cut lists.